backend/requirements.txt
requests==2.27.1
django==3.2.13
djangorestframework==3.13.1
django-filter==21.1
djangorestframework-simplejwt==5.1.0
pillow==9.1.1
drf-extra-fields==3.4.0